add e2e test in readme and actioin (#109)
* add e2e test in readme and actioin


* swfix permission not init


---------

+++ common/infrastructure/postgresql/db.go
@@ -24,7 +24,7 @@ func Init(cfg *Config, removeCfg bool) error {
 		return err
 	}
 
-	if removeCfg {
+	if removeCfg && cfg.Dbcert != "" {
 		if err := os.Remove(cfg.Dbcert); err != nil {
 			return err
 		}
